+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Clever Client changelog
 
+## Unreleased (????-??-??)
+
+* Add `execWarpscript()` function for browsers with fetch
+
 ## 7.2.0 (2021-01-26)
 
 * Add new addon-providers API `GET /v4/addon-providers/{providerId}` in `api/v4/addon-providers.js`